+++ CyRK/nb/nbrk.py
@@ -1,4 +1,5 @@
 from typing import Tuple
+import platform
 
 import numpy as np
 from numba import njit
@@ -83,7 +84,7 @@ def _norm(x):
     return np.linalg.norm(x) / np.sqrt(x.size)
 
 
-@njit(cache=False, fastmath=True)
+@njit(cache=False, fastmath=False)
 def nbrk_ode(
         diffeq: callable, t_span: Tuple[float, float], y0: np.ndarray, args: tuple = tuple(),
         rtol: float = 1.e-6, atol: float = 1.e-8,
@@ -321,7 +322,6 @@ def nbrk_ode(
 
     # Find first step size
     first_step_found = False
-    min_step = EPS_10
     if first_step is not None:
         step_size = max_step
         if first_step < 0.:
@@ -381,7 +381,8 @@ def nbrk_ode(
             else:
                 h1 = (0.01 / max(d1, d2))**error_expo
 
-            step_size = max(min_step, min(100. * h0, h1))
+            next_after = 10. * abs(np.nextafter(t_old, direction * np.inf) - t_old)
+            step_size  = max(next_after, min(100. * h0, h1))
 
     # Main integration loop
     # # Time Loop
@@ -395,6 +396,10 @@ def nbrk_ode(
 
         # Run RK integration step
         # Determine step size based on previous loop
+        # Find minimum step size based on the value of t (less floating point numbers between numbers when t is large)
+        next_after = 10. * abs(np.nextafter(t_old, direction * np.inf) - t_old)
+        min_step   = next_after
+
         # Look for over/undershoots in previous step size
         if step_size > max_step:
             step_size = max_step